requirements.txt
streamlit
pymupdf
python-dotenv
requests
pydantic
networkx
